requirements.txt
streamlit==1.31.1
torch==2.2.0
transformers==4.37.2
pillow==10.2.0
regex==2023.12.25